@@ -1,8 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Web;
-using System.Web.Mvc;
+﻿using System.Web.Mvc;
 
 namespace NgocDuPrint.Controllers
 {
@@ -63,6 +59,10 @@ namespace NgocDuPrint.Controllers
             return View();
         }
         public ActionResult InToRoi()
+        {
+            return View();
+        }
+        public ActionResult BangGia()
         {
             return View();
         }
